requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.2
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pycparser==2.22
sqlparse==0.5.3
tzdata==2025.2
